[react-native-charts-wrapper] Fix test
Because the array of limit lines is built dynamically, there is no way
for older TS versions to know that the type of `lineColor` should remain
the specific unique symbol, but is instead widened to any symbol.

Explicitly typing the object literal fixes this, in similar way that literals
can have their specificity preserved with `as const`.